+++ CHANGELOG.md
@@ -10,6 +10,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - Windows Terminal displays colors fine. We can now remove the win32 workaround. (issue #3779)
   - On the `Google Fonts` profile, the lists of exceptions for **Reserved Font Names (RFN)** and **CamelCased family names**, are now placed on separate txt files (`Lib/fontbakery/data/googlefonts/*_exceptions.txt`) to facilitate their future editing. (issue #3707)
   - The FontVal checks report will be written to a temporary directory now, making it safe to run the checks in parallel on multiple fonts.
+  - Updated the Google Fonts metadata proto format.
 
 ### BugFixes
   - Users reading markdown reports are now directed to the "stable" version of our ReadTheDocs documentation instead of the "latest" (git dev) one. (issue #3677)